Code/requirements.txt
backcall==0.1.0
bleach==3.0.2
certifi==2018.8.24
chardet==3.0.4
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
docopt==0.6.2
entrypoints==0.2.3
findspark==1.3.0
future==0.16.0
happybase==1.1.0
hbase-python==0.5
hdfs==2.1.0
idna==2.7
ipykernel==5.1.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kafka-python==1.4.3
kazoo==2.6.0
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==3.0.0
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
nltk==3.3
notebook==5.7.0
numpy==1.15.2
oauthlib==2.1.0
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
ply==3.11
prometheus-client==0.4.2
prompt-toolkit==2.0.6
protobuf==3.6.1
ptyprocess==0.6.0
py4j==0.10.7
Pygments==2.2.0
PyHive==0.6.1
pyparsing==2.2.2
PySocks==1.6.8
pyspark==2.4.0
python-dateutil==2.7.3
python-twitter==3.4.2
pytz==2018.5
pyzmq==17.1.2
qtconsole==4.4.2
requests==2.19.1
requests-oauthlib==1.0.0
sasl==0.2.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
snakebite==2.11.0
terminado==0.8.1
testpath==0.4.2
textblob==0.15.1
thrift==0.11.0
thrift-sasl==0.3.0
thriftpy==0.3.9
tornado==5.1.1
traitlets==4.3.2
tweepy==3.6.0
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
wordcloud==1.5.0
